Finished lab.

+++ labs/lab-07/handout/playlist_period.cpp
@@ -1,5 +1,5 @@
 #include "playlist_period.h"
 
-size_t playlist_period(const Playlist& playlist) {
+uint32_t playlist_period(const Playlist &playlist) {
     // TODO: Return the period of the playlist.
 }
